Issue #55 パスの整理
