--- frontend/src/app/daily-record/page.tsx
@@ -1,58 +1,0 @@
-"use client";
-
-import * as React from "react";
-
-import { Calendar, BookOpen } from "lucide-react";
-
-import Header from "@/components/common/container/header";
-import { DailyRecordTable } from "@/components/pages/daily-record/daily-record-table";
-import { Button } from "@/components/ui/button";
-import { HeaderCard } from "@/components/ui/header-card";
-import { SectionCard } from "@/components/ui/section-card";
-
-import { useDailyRecord } from "./hooks";
-
-export const DailyRecordPage: React.FC = () => {
-  const { date, rule } = useDailyRecord();
-
-  return (
-    <div className="flex-1 bg-white min-h-screen font-jp">
-      <Header />
-
-      <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8 py-8">
-        <section className="mb-6">
-          <HeaderCard title="本日の成績">
-            <span className="inline-flex items-center gap-1">
-              <Calendar size={14} className="text-white" />
-              <span>{date}</span>
-            </span>
-
-            {rule && (
-              <span className="inline-flex items-center gap-1">
-                <BookOpen size={14} className="text-white" />
-                <span>{rule}</span>
-              </span>
-            )}
-          </HeaderCard>
-        </section>
-
-        <section className="mb-10">
-          <SectionCard title="成績表" bodyClassName="p-0 overflow-x-auto">
-            <DailyRecordTable />
-          </SectionCard>
-        </section>
-
-        <section className="mb-6">
-          <div className="flex flex-col items-center gap-3">
-            <Button variant="brand-secondary" className="w-40 flex-none">
-              記録を追加
-            </Button>
-            <Button className="w-40 flex-none">記録を終了</Button>
-          </div>
-        </section>
-      </div>
-    </div>
-  );
-};
-
-export default DailyRecordPage;
