+++ frontend/src/app/login/hooks/index.ts
@@ -1,7 +1,9 @@
 "use client";
 
 import * as React from "react";
+
 import { useRouter } from "next/navigation";
+
 import { loginToApp } from "@/lib/auth/flows";
 
 export const useLoginPage = () => {
@@ -17,9 +19,13 @@ export const useLoginPage = () => {
 
     try {
       await loginToApp({ email, password });
-      router.replace("/home");
+      router.replace("/");
     } catch (submitError) {
-      setError(submitError instanceof Error ? submitError.message : "ログインに失敗しました");
+      setError(
+        submitError instanceof Error
+          ? submitError.message
+          : "ログインに失敗しました"
+      );
     } finally {
       setIsSubmitting(false);
     }
