+++ frontend/src/app/stats/hooks/index.ts
