+++ frontend/src/app/stats/page.tsx
