+++ frontend/src/middleware.ts
@@ -1,10 +1,12 @@
 import type { NextRequest } from "next/server";
 import { NextResponse } from "next/server";
+
 import { SESSION_COOKIE_NAME } from "@/lib/auth/session";
 
-const isProtectedPath = (pathname: string) => pathname.startsWith("/home");
+const isProtectedPath = (pathname: string) => pathname === "/";
 
-const isGuestOnlyPath = (pathname: string) => pathname === "/login" || pathname === "/signup";
+const isGuestOnlyPath = (pathname: string) =>
+  pathname === "/login" || pathname === "/signup";
 
 export const middleware = (request: NextRequest) => {
   const sessionCookie = request.cookies.get(SESSION_COOKIE_NAME)?.value;
@@ -15,12 +17,12 @@ export const middleware = (request: NextRequest) => {
   }
 
   if (sessionCookie && isGuestOnlyPath(pathname)) {
-    return NextResponse.redirect(new URL("/home", request.url));
+    return NextResponse.redirect(new URL("/", request.url));
   }
 
   return NextResponse.next();
 };
 
 export const config = {
-  matcher: ["/home/:path*", "/login", "/signup"],
+  matcher: ["/", "/login", "/signup"],
 };
